遗留的js代码全部转移到源码区内
