+++ src/globals.js
@@ -1,0 +1,27 @@
+// global variables
+// var shishen_json
+var ownteam = []
+var enemyteam = []
+var ban = []
+var historynum = 0
+var queryhistory = {}
+var sp_opts = ''
+var ssr_opts = ''
+var sr_opts = ''
+var r_opts = ''
+var n_opts = ''
+
+let globals = {
+    ownteam,
+    enemyteam,
+    ban,
+    historynum,
+    queryhistory,
+    sp_opts,
+    ssr_opts,
+    sr_opts,
+    r_opts,
+    n_opts
+};
+
+export default globals;
